refactor(gateway): extract GatewaySessionMixin from run.py (slice 21 of #54962)
